application structure WIP
