+++ typademic/static/favicon-192x192.png
