+++ typademic/static/favicon-32x32.png
