+++ typademic/static/manifest.json
