+++ typademic/templates/base.html
@@ -71,26 +71,26 @@
 
     <!-- favicons -->
     <link rel="apple-touch-icon-precomposed" sizes="57x57"
-          href="/static/apple-touch-icon-57x57.png"/>
+          href="/typademic/static/apple-touch-icon-57x57.png"/>
     <link rel="apple-touch-icon-precomposed" sizes="114x114"
-          href="/static/apple-touch-icon-114x114.png"/>
+          href="/typademic/static/apple-touch-icon-114x114.png"/>
     <link rel="apple-touch-icon-precomposed" sizes="72x72"
-          href="/static/apple-touch-icon-72x72.png"/>
+          href="/typademic/static/apple-touch-icon-72x72.png"/>
     <link rel="apple-touch-icon-precomposed" sizes="144x144"
-          href="/static/apple-touch-icon-144x144.png"/>
+          href="/typademic/static/apple-touch-icon-144x144.png"/>
     <link rel="apple-touch-icon-precomposed" sizes="60x60"
-          href="/static/apple-touch-icon-60x60.png"/>
+          href="/typademic/static/apple-touch-icon-60x60.png"/>
     <link rel="apple-touch-icon-precomposed" sizes="120x120"
-          href="/static/apple-touch-icon-120x120.png"/>
+          href="/typademic/static/apple-touch-icon-120x120.png"/>
     <link rel="apple-touch-icon-precomposed" sizes="76x76"
-          href="/static/apple-touch-icon-76x76.png"/>
+          href="/typademic/static/apple-touch-icon-76x76.png"/>
     <link rel="apple-touch-icon-precomposed" sizes="152x152"
-          href="/static/apple-touch-icon-152x152.png"/>
-    <link rel="icon" type="image/png" href="/static/favicon-196x196.png" sizes="196x196"/>
-    <link rel="icon" type="image/png" href="/static/favicon-96x96.png" sizes="96x96"/>
-    <link rel="icon" type="image/png" href="/static/favicon-32x32.png" sizes="32x32"/>
-    <link rel="icon" type="image/png" href="/static/favicon-16x16.png" sizes="16x16"/>
-    <link rel="icon" type="image/png" href="/static/favicon-128.png" sizes="128x128"/>
+          href="/typademic/static/apple-touch-icon-152x152.png"/>
+    <link rel="icon" type="image/png" href="/typademic/static/favicon-196x196.png" sizes="196x196"/>
+    <link rel="icon" type="image/png" href="/typademic/static/favicon-96x96.png" sizes="96x96"/>
+    <link rel="icon" type="image/png" href="/typademic/static/favicon-32x32.png" sizes="32x32"/>
+    <link rel="icon" type="image/png" href="/typademic/static/favicon-16x16.png" sizes="16x16"/>
+    <link rel="icon" type="image/png" href="/typademic/static/favicon-128.png" sizes="128x128"/>
     <meta name="application-name" content="typademic.ch"/>
     <meta name="msapplication-TileColor" content="#FFFFFF"/>
     <meta name="msapplication-TileImage" content="/static/mstile-144x144.png"/>
@@ -98,13 +98,13 @@
     <meta name="msapplication-square150x150logo" content="/static/mstile-150x150.png"/>
     <meta name="msapplication-wide310x150logo" content="/static/mstile-310x150.png"/>
     <meta name="msapplication-square310x310logo" content="/static/mstile-310x310.png"/>
-    <link rel="manifest" href="/static//manifest.json">
+    <link rel="manifest" href="/typademic/static//manifest.json">
     <meta name="apple-mobile-web-app-title" content="typademic.ch">
     <meta name="application-name" content="typademic.ch">
     <meta name="theme-color" content="#ffffff">
     
     <!-- JS -->
-    <script defer src="https://use.fontawesome.com/releases/v5.1.0/js/all.js"></script>
+    <script defer src="https://use.fontawesome.com/releases/v5.3.1/js/all.js"></script>
     <script async src="https://www.googletagmanager.com/gtag/js?id={{ config['GOOGLE_ANALYTICS'] }}"></script>
     <script>
         window.dataLayer = window.dataLayer || [];
